Track/requirements.txt
Django==5.0.7
requests==2.26.0
psycopg2>=2.8
channels
daphne
djangorestframework
django-cors-headers
PyJWT==2.10.0